+++ config/redis.ts
@@ -1,13 +1,13 @@
 const config: Record<string, any> = {
-    default: process.env.REDIS_CONNECTION,
+    default: Bun.env.REDIS_CONNECTION,
 
     connections: {
         local: {
-            host: process.env.REDIS_HOST,
-            port: process.env.REDIS_PORT,
-            password: process.env.REDIS_PASSWORD,
-            database: process.env.REDIS_DATABASE,
-            maxRetries: Number(process.env.REDIS_MAX_RETRIES)
+            host: Bun.env.REDIS_HOST,
+            port: Bun.env.REDIS_PORT,
+            password: Bun.env.REDIS_PASSWORD,
+            database: Bun.env.REDIS_DATABASE,
+            maxRetries: Number(Bun.env.REDIS_MAX_RETRIES)
         }
     }
 };