+++ routes/api.ts
@@ -1,6 +1,6 @@
 import Router from "@bejibun/core/facades/Router";
-import hello from "@/routes/hello";
-import test from "@/routes/test";
+import hello from "@/routes/api/hello";
+import test from "@/routes/api/test";
 
 export default Router.prefix("api").group([
     hello,