+++ routes/web.ts
@@ -1,0 +1,5 @@
+import Router from "@bejibun/core/facades/Router";
+
+export default Router.prefix("web").group([
+    //
+]);